+++ cpp/core/math/ZSH.h
@@ -16,7 +16,9 @@
 
 #pragma once
 
+#include "eigen_plugins/eigen_plugins.h"
 #include <Eigen/Dense>
+#include <cstddef>
 
 #include "math/SH.h"
 #include "math/least_squares.h"
@@ -38,7 +40,7 @@ inline size_t NforL(int lmax) { return (1 + lmax / 2); }
 inline size_t index(int l) { return (l / 2); }
 
 //! returns the largest \e lmax given \a N parameters
-inline size_t LforN(int N) { return (2 * (N - 1)); }
+inline size_t LforN(int N) { return (static_cast<size_t>(2 * (N - 1))); }
 
 //! form the ZSH->amplitudes matrix for a set of inclination angles
 /*! This computes the matrix \a ZSHT mapping zonal spherical harmonic
